Twitter_FactomV2/Chain_Creator/requirements.txt
factom-api==0.1.1
ipython
dataset
textblob
psycopg2>=2.7,<3.0